Fix up the windows build.
Former-commit-id: 386255d2016e49c3ec7696dfc4cf0bed11f3ae79

+++ libdevcrypto/Common.cpp
@@ -345,9 +345,9 @@ void Nonce::initialiseIfNeeded()
 	{
 		// todo: replace w/entropy from user and system
 		std::mt19937_64 s_eng(time(0) + chrono::high_resolution_clock::now().time_since_epoch().count());
-		std::uniform_int_distribution<uint8_t> d(0, 255);
+		std::uniform_int_distribution<uint16_t> d(0, 255);
 		for (unsigned i = 0; i < 32; ++i)
-			m_value[i] = d(s_eng);
+			m_value[i] = (uint8_t)d(s_eng);
 	}
 	if (!m_value)
 		BOOST_THROW_EXCEPTION(InvalidState());
